refactor(runtime): merge openai_responses into openai provider
Fold the Responses API path into OpenAIProvider, routed per-model by
usesResponsesApi() (gated on the real openai provider so compatible
subclasses stay on Chat Completions). Retire the openai_responses
provider id.

Also prune pre-gpt-5 OpenAI models: the openai provider now serves only
the gpt-5 family; the live model list is filtered and the fallback list
and recommended anchor updated accordingly.
